1/10 7:37 Main World and Menu Updates
Updated Main world with landscape painting. Might change later since it still looks pixelated.
Added the controls UI and added functionality to the Pause menu with resume, controls, and quit game, but need to add options and credits.
Finished Mayor Dialogue, but need to add the checks, if to be implemented (otherwise mayor dialogues are finished).
Added Deer, and the deer antler trees with special VFX to them