Life, Death, and Lost Loot
Here are the changes as follows:
Health System with a UI Health Bar
For anything in the game you want to damage the player, in its blueprint cast to the player and call TakeDamage and then assign a Damage Amount! The rest will handle itself.
Death System with its own UI
Respawning System inspired by Dark Souls
There is a BP_Shrine now, this works like the other placed interactables in the Snail Base where an interaction prompt will appear when the player approaches close enough.
When interacted, the shrine will clear the player's previously saved spawn point (defaulted to the original player start), and when you die and respawn, you will spawn just in front of the shrine.
Implemented the Lost Inventory solution for Respawning
When the player dies, a bag drops on the ground. This copies the player's inventory and adds it to the Lost Inventory.
When the player walks into the bag, the player reclaims the lost inventory. This works even when there are items already present in the inventory.
If a player were to die before reclaiming the lost inventory, the bag will despawn.
I've tested it through many test cases and I believe I have roughed through the kinks. Further testing will be done. It is functional and feels right. Pushing now.