Added the BP_GeyserRock and the BP_GeyserTrigger
The details panel of the BP_GeyserRock has multipliers that will adjust how intense the rumble animation script, the height script, and the bobbing script are. The BP_GeyserTrigger has an array in the details panel that you can make as big as you need for all the geysers you want to trigger from the trigger box.  In the level Armoni_Practice_Level_MainWorld2 has the example